add test for slash handler -- /hello test passes
